% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtmosphericEmissivity.R
\name{AtmosphericEmissivity}
\alias{AtmosphericEmissivity}
\title{AtmosphericEmissivity}
\usage{
AtmosphericEmissivity(airtemp, cloudiness, vp = NULL, opt = "linear")
}
\arguments{
\item{airtemp}{Air temperature in degree Celcius}

\item{cloudiness}{Cloudiness in fraction}

\item{vp}{Vapor pressure}

\item{opt}{option (default=linear)}
}
\value{
Atmospheric Emissivity
}
\description{
AtmosphericEmissivity
}
