% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declination.R
\name{declination}
\alias{declination}
\title{solar declination}
\usage{
declination(Jday)
}
\arguments{
\item{Jday}{Julian date or day of the year [day]}
}
\value{
solar declination [rad]
}
\description{
solar declination
}
