% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstCloudiness.R
\name{EstCloudiness}
\alias{EstCloudiness}
\title{Cloudiness}
\usage{
EstCloudiness(
  Tx = (-999),
  Tn = (-999),
  trans = NULL,
  transMin = 0.15,
  transMax = 0.75,
  opt = "linear"
)
}
\arguments{
\item{Tx}{Daily Maximum temperature in degree Celcius}

\item{Tn}{Daily Minimum temperature in degree Celcius}

\item{trans}{transmissivity (default=NULL--> use transmissivity function)}

\item{transMin}{Minimum transmissivity (default=0.15)}

\item{transMax}{Maximum transmissivity (default=0.75)}

\item{opt}{Cloudiness estimation methods: Black or linear (default=linear)}
}
\value{
Cloudiness in fraction
}
\description{
Cloudiness
}
