% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Longwave.R
\name{Longwave}
\alias{Longwave}
\title{Longwave}
\usage{
Longwave(emissivity, temp, TimeStep = 24)
}
\arguments{
\item{emissivity}{Terrestrial Emissivity}

\item{temp}{temperature of the emitting body in degree Celsius}

\item{TimeStep}{Time step in hour: Ex) 1 hour=1, daily=24 (default)}
}
\value{
Incoming longwave radiation [kJ m-2/TimeStep]
}
\description{
Incoming longwave radiation based on the Sephan-Boltzman equation
}
