% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PotentialSolar.R
\name{PotentialSolar}
\alias{PotentialSolar}
\title{Potential Solar}
\usage{
PotentialSolar(lat, Jday)
}
\arguments{
\item{Jday}{}
}
\value{
Extraterrestrial (potential) solar radiation [kJ m-2 d-1]
}
\description{
Extraterrestrial (potential) solar radiation at the edge of the atmospher
}
