% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SatVaporDensity.R
\name{SatVaporDensity}
\alias{SatVaporDensity}
\title{Saturated Vapor Density}
\usage{
SatVaporDensity(T_C)
}
\arguments{
\item{T_C}{Temperature in degree Celsius}
}
\value{
Saturated Vapor Density [kg/m3]
}
\description{
Saturated Vapor Density
}
\examples{
SatVaporDensity(5)
}
