% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopefactor.R
\name{slopefactor}
\alias{slopefactor}
\title{Slope factor}
\usage{
slopefactor(lat, Jday, slope, aspect)
}
\arguments{
\item{lat}{Latitde [rad]}

\item{Jday}{Julian date or day of the year [day]}

\item{slope}{Slope of the ground [rad]}

\item{aspect}{Ground aspect [rad from north]}
}
\value{
Slope factor [0,1]
}
\description{
Adjust solar radiation for land slope and aspect relative to the sun, 1=level ground
}
