% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnowMelt_EPA.R
\name{SnowMelt_EPA}
\alias{SnowMelt_EPA}
\title{EPA PBSM snow model}
\usage{
SnowMelt_EPA(
  Date,
  precip_mm,
  Tmax_C,
  Tmin_C,
  windSp_ms,
  Relhum_frac,
  swrad_wm2,
  lwrad_wm2,
  lat_deg,
  slope = 0,
  aspect = 0,
  tempHt = 1,
  windHt = 2,
  groundAlbedo = 0.25,
  SurfEmissiv = 0.95,
  forest = 0,
  startingSnowDepth_m = 0,
  startingSnowDensity_kg_m3 = 550,
  Train = 10,
  Tsnow = 0,
  Max_SnowDensity_kg_m3 = 550
)
}
\arguments{
\item{Date}{Date(YYYY-MM-DD)}

\item{precip_mm}{Precipitation in mm/day}

\item{Tmax_C}{Maximum temperature in degree Celsius}

\item{Tmin_C}{Minimum temperature in degree Celsius}

\item{windSp_ms}{Wind speed in m/sec}

\item{Relhum_frac}{Relative humidity in fraction}

\item{swrad_wm2}{Incoming shortwave (solar) radiation in W/m2}

\item{lwrad_wm2}{Incoming longwave radiation in W/m2}

\item{lat_deg}{Latitude in degree}

\item{slope}{Slope factor}

\item{aspect}{Ground aspect [rad from north]}

\item{tempHt}{Height of temperature station}

\item{windHt}{Height of wind station}

\item{groundAlbedo}{Ground albedo (default=0.25)}

\item{SurfEmissiv}{Surface Emissivity}

\item{forest}{Forest ratio}

\item{startingSnowDepth_m}{Initial value of snow depth in meter}

\item{startingSnowDensity_kg_m3}{Initial value of snow density in kg/m3}

\item{Train}{Upper temperature of S curve: precipitation is all rain if temperature >= Train}

\item{Tsnow}{Lower temperature of S curve: precipitation is all snow if temperature <= Tsnow}

\item{Max_SnowDensity_kg_m3}{Maximum snow density in kg/m3, default=550 from Voordendag et al.(2021)}
}
\value{
Time series of snow simulation
}
\description{
Process-based snow model with seven hydroclimate variables (prcp, min/max temperature, wind speed, relative humidity and short/longwave radiations)
}
\details{
A modified version of PBSM-S-Curve which simulates snow processes with seven hydroclimate variables. All of hydroclimate variables are from new version of the hybrid climate dataset.
}
