% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnowMelt_EPA_Subdaily.R
\name{SnowMelt_EPA_Subdaily}
\alias{SnowMelt_EPA_Subdaily}
\title{EPA PBSM snow model}
\usage{
SnowMelt_EPA_Subdaily(
  Date,
  time.step = 1,
  precip_mm,
  Tair,
  windSp_ms,
  Relhum_frac,
  swrad_wm2,
  lwrad_wm2,
  lat_deg,
  tempHt = 1,
  windHt = 2,
  groundAlbedo = 0.25,
  SurfEmissiv = 0.95,
  forest = 0,
  startingSnowDepth_m = 0,
  startingSnowDensity_kg_m3 = 550,
  Max_SnowDensity_kg_m3 = 550
)
}
\arguments{
\item{Date}{Date(YYYY-MM-DD HH:MM)}

\item{time.step}{Time.step in hours (default=1)}

\item{precip_mm}{Precipitation in mm/day}

\item{Tair}{Air temperature in degree Celsius}

\item{windSp_ms}{Wind speed in m/sec}

\item{Relhum_frac}{Relative humidity in fraction}

\item{swrad_wm2}{Incoming shortwave (solar) radiation in W/m2}

\item{lwrad_wm2}{Incoming longwave radiation in W/m2}

\item{lat_deg}{Latitude in degree}

\item{tempHt}{Height of temperature station}

\item{windHt}{Height of wind station}

\item{groundAlbedo}{Ground albedo (default=0.25)}

\item{SurfEmissiv}{Surface Emissivity}

\item{forest}{Forest ratio}

\item{startingSnowDepth_m}{Initial value of snow depth in meter}

\item{startingSnowDensity_kg_m3}{Initial value of snow density in kg/m3}

\item{Max_SnowDensity_kg_m3}{Maximum snow density in kg/m3, default=550 from Voordendag et al.(2021)}
}
\value{
Time series of snow simulation
}
\description{
Process-based snow model with six hydroclimate variables (prcp, air temperature, wind speed, relative humidity and short/longwave radiations) at a sub-daily time step (e.g., hourly data)
}
\details{
A modified version of PBSM which simulates snow processes with seven hydroclimate variables at sub-daily time step. All of hydroclimate variables are usually from weather forecast data (e.g., GEPS and CFS).
}
