% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnowMelt_S_Curve.R
\name{SnowMelt_S_Curve}
\alias{SnowMelt_S_Curve}
\title{PBSM with S-curve}
\usage{
SnowMelt_S_Curve(
  Date,
  precip_mm,
  Tmax_C,
  Tmin_C,
  lat_deg,
  slope = 0,
  aspect = 0,
  tempHt = 1,
  windHt = 2,
  groundAlbedo = 0.25,
  SurfEmissiv = 0.95,
  windSp = 2,
  forest = 0,
  startingSnowDepth_m = 0,
  startingSnowDensity_kg_m3 = 550,
  Train = 10,
  Tsnow = 0,
  Max_SnowDensity_kg_m3 = 550
)
}
\arguments{
\item{Date}{Date (YYYY-MM-DD)}

\item{precip_mm}{Precipitation in mm/day}

\item{Tmax_C}{Maximum temperature in degree Celsius}

\item{Tmin_C}{Minimum temperature in degree Celsius}

\item{lat_deg}{Latitude in degree}

\item{slope}{Slope factor}

\item{aspect}{Ground aspect [rad from north]}

\item{tempHt}{Height of temperature station}

\item{windHt}{Height of wind station}

\item{groundAlbedo}{Ground albedo (default=0.25)}

\item{SurfEmissiv}{Surface Emissivity}

\item{windSp}{Wind speed in m/sec (default=2.0)}

\item{forest}{Forest ratio}

\item{startingSnowDepth_m}{Initial value of snow depth in meter}

\item{startingSnowDensity_kg_m3}{Initial value of snow density in kg/m3}

\item{Train}{Upper temperature of S curve: precipitation is all rain if temperature >= Train}

\item{Tsnow}{Lower temperature of S curve: precipitation is all snow if temperature <= Tsnow}

\item{Max_SnowDensity_kg_m3}{Maximum snow density in kg/m3, default=550 from Voordendag et al.(2021)}
}
\value{
Time series of snow simulation
}
\description{
Original PBSM has been Modified by incorporating a S-shape curve method that separates rain and snow by a S-curve
Added ice rain in snowpack if snowpack temperature is below 0 and rain is on the snowpack.
}
\details{
Added two arguments (Train and Tsnow) that are minimum temperature (Train) for all precipitation and maximum temperature (Tsnow) for all snow.
Author: Hyung Eum, AEP
Date: March 2021
}
\references{
Todd Walter, M., Brooks, E.S., McCool, D.K., King, L.G., Molnau, M., Boll, J., 2005. Process-based snowmelt modeling: does it require more input data than temperature-index modeling? Journal of Hydrology 300, 65–75. https://doi.org/10.1016/j.jhydrol.2004.05.002

Fuka, D.R., Easton, Z.M., Brooks, E.S., Boll, J., Steenhuis, T.S., Walter, M.T., 2012. A Simple Process-Based Snowmelt Routine to Model Spatially Distributed Snow Depth and Snowmelt in the SWAT Model. JAWRA Journal of the American Water Resources Association 48, 1151–1161. https://doi.org/10.1111/j.1752-1688.2012.00680.x

Stefan W. Kienzle (2008). A new temperature based method to separate rain and snow. Hydrological processes, 22, 5067-5085.
}
