% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnowMelt_S_Curve_Sublim.R
\name{SnowMelt_S_Curve_Sublim}
\alias{SnowMelt_S_Curve_Sublim}
\title{PBSM with S-curve and Sublimation}
\usage{
SnowMelt_S_Curve_Sublim(
  Date,
  precip_mm,
  Tmax_C,
  Tmin_C,
  lat_deg,
  slope = 0,
  aspect = 0,
  tempHt = 1,
  windHt = 2,
  groundAlbedo = 0.25,
  SurfEmissiv = 0.95,
  windSp = 2,
  forest = 0,
  startingSnowDepth_m = 0,
  startingSnowDensity_kg_m3 = 550,
  Train = 10,
  Tsnow = 0,
  Max_SnowDensity_kg_m3 = 550
)
}
\arguments{
\item{Date}{Date (YYYY-MM-DD)}

\item{precip_mm}{Precipitation in mm/day}

\item{Tmax_C}{Maximum temperature in degree Celsius}

\item{Tmin_C}{Minimum temperature in degree Celsius}

\item{lat_deg}{Latitude in degree}

\item{slope}{Slope factor}

\item{aspect}{Ground aspect [rad from north]}

\item{tempHt}{Height of temperature station}

\item{windHt}{Height of wind station}

\item{groundAlbedo}{Ground albedo (default=0.25)}

\item{SurfEmissiv}{Surface Emissivity}

\item{windSp}{Wind speed in m/sec (default=2.0)}

\item{forest}{Forest ratio}

\item{startingSnowDepth_m}{Initial value of snow depth in meter}

\item{startingSnowDensity_kg_m3}{Initial value of snow density in kg/m3}

\item{Train}{Upper temperature of S curve: precipitation is all rain if temperature >= Train}

\item{Tsnow}{Lower temperature of S curve: precipitation is all snow if temperature <= Tsnow}

\item{Max_SnowDensity_kg_m3}{Maximum snow density in kg/m3, default=550 from Voordendag et al.(2021)}
}
\value{
Time series of snow simulation including sublimation
}
\description{
Original PBSM has been Modified by incorporating a S-shape curve method that separates rain and snow by a S-curve
Added ice rain in snowpack if snowpack temperature is below 0 and rain is on the snowpack.
Added Sublimation physics (Mass transfer based on Latent Heat Flux).
}
\details{
Added two arguments (Train and Tsnow) that are minimum temperature (Train) for all precipitation and maximum temperature (Tsnow) for all snow.
Author: Hyung Eum, AEP
Date: March 2021 (Modified for Sublimation Feb 2026)
}
