% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solar.R
\name{Solar}
\alias{Solar}
\title{Solar radiation}
\usage{
Solar(
  lat,
  Jday,
  Tx,
  Tn,
  albedo = 0.2,
  forest = 0,
  slope = 0,
  aspect = 0,
  units = "kJm2d",
  latUnits = "unknown",
  printWarn = TRUE
)
}
\arguments{
\item{lat}{Latitude: degree or rad}

\item{Jday}{Julian date or day of the year [day]}

\item{Tx}{Maximum temperature in degree Celcius}

\item{Tn}{Minimum temperature in degree Celcius}

\item{albedo}{Surface albedo (default=0.2)}

\item{forest}{Forest ratio}

\item{slope}{Slope}

\item{aspect}{Ground aspect [rad from north]}

\item{units}{Radiation unit: Wm2 or kJm2d (default=kJm2d)}

\item{latUnits}{Latitude unit (default=unknown)}

\item{printWarn}{option for printing warning (default=TRUE)}
}
\value{
Estimated solar radiation
}
\description{
Estimate solar radiation
}
