% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarangle.R
\name{solarangle}
\alias{solarangle}
\title{Solar Angle}
\usage{
solarangle(lat, Jday)
}
\arguments{
\item{lat}{latitdue [rad]}

\item{Jday}{Julian date or day of the year [day]}
}
\value{
Solar angle in radian
}
\description{
Angle of solar inclination from horizontal at solar noon [rad]
}
