% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transmissivity.R
\name{transmissivity}
\alias{transmissivity}
\title{Transmissivity}
\usage{
transmissivity(Tx, Tn, A = 0.75, C = 2.4, opt = "1day", JD = NULL)
}
\arguments{
\item{Tx}{maximum daily temperature [C]}

\item{Tn}{minimum daily temperature [C]}

\item{A}{Empirical coefficient (default=0.75)}

\item{C}{Empirical coefficient (default=2.4)}

\item{opt}{average diurnal temperature methods: 1day (default), 2day and missingdays}

\item{JD}{Julian Day (default=NULL)}
}
\value{
Atmospheric transmissivity
}
\description{
Fraction of direct solar radiation passing through the atmosphere based on the Bristow-Campbell eqn
}
